feat: Add Annotations Support
Separates the note creation into two workflows: bookmarks and annotations. The presence of the `annotations` property is used as the flag to decide which function to execute. Within the annotation function, true bookmarks get a special syntax "Page bookmark" to distinguish them better from highlights and notes.

This code can probably be optimized further.

Future feature improvements may be to utilize the annotation's `color` property to set the block's color.